+++ storage_utils/experimental/s3.py
@@ -1,7 +1,7 @@
 """
 An experimental implementation of S3 in storage-utils.
 """
-import posixpath
+import tempfile
 import threading
 
 import boto3
@@ -9,11 +9,14 @@ from botocore import exceptions as boto_s3_exceptions
 
 from storage_utils import exceptions
 from storage_utils.base import Path
-from storage_utils.posix import PosixPath
+from storage_utils.remote import RemoteFile
+from storage_utils.remote import RemotePath
 from storage_utils import utils
 
 # Thread-local variable used to cache the client
 _thread_local = threading.local()
+
+S3File = RemoteFile
 
 
 def _parse_s3_error(exc):
@@ -48,7 +51,7 @@ def _get_s3_client():
     return _thread_local.s3_client
 
 
-class S3Path(Path):
+class S3Path(RemotePath):
     """
     Provides the ability to manipulate and access S3 resources
     with a similar interface to the path library.
@@ -57,61 +60,12 @@ class S3Path(Path):
     near-future, users should be able to custom configure the S3 client.
     """
     drive = 's3://'
-    path_module = posixpath
-    parts_class = PosixPath
-
-    def __init__(self, pth):
-        """
-        Validates S3 path is in the proper format.
-
-        Args:
-            pth (str): A path that matches the format of
-                ``s3://{bucket_name}/{rest_of_path}``
-                The ``s3://`` prefix is required in the path.
-        """
-        if not hasattr(pth, 'startswith') or not pth.startswith(self.drive):
-            raise ValueError('path must have %s (got %r)' % (self.drive, pth))
-        return super(S3Path, self).__init__(pth)
-
-    def __repr__(self):
-        return '%s("%s")' % (type(self).__name__, self)
-
-    def _get_parts(self):
-        """Returns the path parts (excluding s3://) as a list of strings."""
-        if len(self) > len(self.drive):
-            return self[len(self.drive):].split('/')
-        else:
-            return []
-
-    @property
-    def name(self):
-        """The name of the path, mimicking path.py's name property"""
-        return self.parts_class(super(S3Path, self).name)
-
-    @property
-    def parent(self):
-        """The parent of the path, mimicking path.py's parent property"""
-        return self.path_class(super(S3Path, self).parent)
 
     @property
     def bucket(self):
         """Returns the bucket name from the path or None"""
         parts = self._get_parts()
         return parts[0] if len(parts) > 0 and parts[0] else None
-
-    @property
-    def resource(self):
-        """
-        Returns the resource as a ``PosixPath`` object or None.
-
-        A resource can be a single object or a prefix to objects.
-        Note that it's important to keep the trailing slash in a resource
-        name for prefix queries.
-        """
-        parts = self._get_parts()
-        joined_resource = '/'.join(parts[1:]) if len(parts) > 1 else None
-
-        return self.parts_class(joined_resource) if joined_resource else None
 
     def _s3_client_call(self, method_name, *args, **kwargs):
         """
@@ -150,7 +104,7 @@ class S3Path(Path):
         Raises:
             RemoteError: A s3 client error occurred.
         """
-        raise NotImplementedError
+        return S3File(self, mode=mode)
 
     def list(self, starts_with=None, limit=None, list_as_dir=False):
         """
@@ -371,20 +325,31 @@ class S3Path(Path):
         del response['ResponseMetadata']
         return response
 
-    def walkfiles(self, pattern=None):
-        """
-        Iterate over listed files whose filenames match an optional pattern.
+    def read_object(self):
+        """"Reads an individual object."""
+        body = self._s3_client_call('get_object', Bucket=self.bucket, Key=self.resource)['Body']
+        return body.read()
+
+    def write_object(self, content):
+        """Writes an individual object.
+
+        Note that this method writes the provided content to a temporary
+        file before uploading.
 
         Args:
-            pattern (str, optional): Only return files that match this pattern.
-
-        Returns:
-            Iter[S3Path]: All files that match the optional pattern. Directories
-                are not returned.
+            content (str): The content of the object
+            **swift_upload_args: Keyword arguments to pass to
+                `SwiftPath.upload`
         """
-        for f in self.list():
-            if pattern is None or f.fnmatch(pattern):
-                yield f
+        with tempfile.NamedTemporaryFile() as fp:
+            fp.write(content)
+            fp.flush()
+            ul_kwargs = {
+                'Bucket': self.bucket,
+                'Key': self.resource,
+                'Filename': fp.name
+            }
+            self._s3_client_call('upload_file', **ul_kwargs)
 
     def download_object(self, dest):
         """
@@ -441,7 +406,6 @@ class S3Path(Path):
           directory.
 
         TODO:
-
         - Update once directory markers are implemented
 
         """