+++ docs/release_notes.rst
@@ -90,7 +90,7 @@ v6.4.0
 v6.3.1
 ------
 
-* Add `SwiftFile.tell` method to enable `SwiftFile` to work with `gzip.GzipFile` objects transparently.
+* Add ``SwiftFile.tell`` method to enable `SwiftFile` to work with `gzip.GzipFile` objects transparently.
 * Make `SwiftFile` delegated methods show up in docs.
 
 v6.3.0